relocated SimplifyType
