+++ packages/libs/sdk/package.json
@@ -8,8 +8,9 @@
   "license": "MIT",
   "version": "1.1.0",
   "main": "./cjs/index.js",
-  "module": "./esm/index.js",
+  "module": "./esm/src/index.js",
   "types": "./index.d.ts",
+  "sideEffects": false,
   "dependencies": {
     "@urql/core": "^4.0.7",
     "@urql/exchange-graphcache": "^6.0.4",
@@ -42,16 +43,16 @@
   },
   "exports": {
     ".": {
-      "import": "./esm/index.js",
+      "import": "./esm/src/index.js",
       "require": "./cjs/index.js"
     },
     "./api": {
-      "import": "./esm/api/index.js",
-      "types": "./esm/index.d.ts"
+      "import": "./esm/src/api/index.js",
+      "types": "./esm/src/api/index.d.ts"
     },
-    "./viem": {
-      "import": "./esm/viemExport/index.js",
-      "types": "./esm/index.d.ts"
+    "./core": {
+      "import": "./esm/src/core/index.js",
+      "types": "./esm/src/core/index.d.ts"
     }
   }
 }
