+++ packages/libs/sdk/src/core/addOns/nftTokenV2/types/qn_getTransactionsByAddress.ts
@@ -1,4 +1,4 @@
-import { type SimplifyType } from '../../../../../src/api/utils/helpers';
+import { type SimplifyType } from '../../../../lib/types';
 import { z } from 'zod';
 import {
   isEvmAddress,
