+++ packages/libs/sdk/src/core/addOns/nftTokenV2/types/qn_verifyNFTsOwner.ts
@@ -1,4 +1,4 @@
-import { type SimplifyType } from '../../../../../src/api/utils/helpers';
+import { type SimplifyType } from '../../../../lib/types';
 import { z } from 'zod';
 import { isEvmAddress } from '../../../../lib/validation/validators';
