+++ packages/libs/sdk/src/lib/types/SimplifyType.ts
@@ -1,0 +1,3 @@
+export type SimplifyType<T> = T extends object
+  ? { [K in keyof T]: SimplifyType<T[K]> }
+  : T;
